Refactor project by removing the Internal (development) IDP. Deleted anything used only by the IDP code.
