+++ saml-handler/src/main/java/org/apache/sling/auth/saml2/TokenStore.java
@@ -32,7 +32,6 @@ import java.security.SecureRandom;
 import javax.crypto.Mac;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
-
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
