+++ saml-handler/src/main/java/org/apache/sling/auth/saml2/sp/Saml2LoginModule.java
@@ -17,15 +17,11 @@
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
  */
 
-package org.apache.sling.auth.saml2.sync;
+package org.apache.sling.auth.saml2.sp;
 
 import org.apache.jackrabbit.oak.spi.security.authentication.AbstractLoginModule;
 import org.apache.jackrabbit.oak.spi.security.authentication.PreAuthenticatedLogin;
-import org.apache.jackrabbit.oak.spi.security.authentication.external.*;
-import org.apache.jackrabbit.oak.spi.whiteboard.Whiteboard;
-import org.apache.jackrabbit.oak.spi.whiteboard.WhiteboardUtils;
 import org.apache.sling.auth.saml2.impl.Saml2Credentials;
-import org.apache.sling.auth.saml2.impl.Saml2ExternalIdentityProvider;
 import org.osgi.service.component.annotations.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -43,15 +39,11 @@ import java.util.Set;
         service = Saml2LoginModule.class,
         immediate = true
 )
-//        name = "SAML2 Login Module",
+
 
 public class Saml2LoginModule extends AbstractLoginModule {
 
     static Set<Class> SUPPORTED_CREDENTIALS = new HashSet<Class>();
-    private static final int MAX_SYNC_ATTEMPTS = 3;
-    private ExternalIdentityProvider externalIdentityProvider;
-    private SyncHandler syncHandler;
-    private ExternalUser externalUser;
     private final Logger logger = LoggerFactory.getLogger(Saml2LoginModule.class);
 
     static {
@@ -95,7 +87,7 @@ public class Saml2LoginModule extends AbstractLoginModule {
                 sharedState.put(SHARED_KEY_PRE_AUTH_LOGIN, new PreAuthenticatedLogin(userId));
                 sharedState.put(SHARED_KEY_CREDENTIALS, new SimpleCredentials(userId, new char[0]));
                 sharedState.put(SHARED_KEY_LOGIN_NAME, userId);
-                logger.info("Adding pre-authenticated login user '{}' to shared state.", userId);
+                logger.debug("Adding pre-authenticated login user '{}' to shared state.", userId);
             }
         }
         return false;
@@ -131,51 +123,12 @@ public class Saml2LoginModule extends AbstractLoginModule {
     @Override
     public void initialize(Subject subject, CallbackHandler callbackHandler, Map<String, ?> sharedState, Map<String, ?> options) {
         super.initialize(subject, callbackHandler, sharedState, options);
-        // 1. Get Whiteboard
-        Whiteboard whiteboard = getWhiteboard();
-        if (null == whiteboard) {
-            logger.error("Header login module needs Whiteboard.");
-            return;
-        }
-        logger.debug("Got Whiteboard");
-
-        // 2. Get Identity Provider Manager
-        ExternalIdentityProviderManager externalIdentityProviderManager = WhiteboardUtils.getService(whiteboard, ExternalIdentityProviderManager.class);
-        if (null == externalIdentityProviderManager) {
-            logger.error("SAML2 login module needs ExternalIdentityProviderManager.");
-            return;
-        }
-
-        // 3. get Identity Provider
-        externalIdentityProvider = externalIdentityProviderManager.getProvider(Saml2ExternalIdentityProvider.NAME);
-        if (null == externalIdentityProvider) {
-            logger.error("Header login module needs ExternalIdentityProvider. Can't get IDP with name: '{}'",
-                    Saml2ExternalIdentityProvider.NAME);
-            return;
-        }
-
-// 4. get Sync Manager
-//        SyncManager syncManager = WhiteboardUtils.getService(whiteboard, SyncManager.class);
-//        if (null == syncManager)
-//        {
-//            logger.error("Header login module needs SyncManager.");
-//            return;
-//        }
-
-// 5. Get Sync Handler
-//        String syncHandlerName = "default";
-//        syncHandler = syncManager.getSyncHandler(syncHandlerName);
-//        if (null == syncHandler)
-//        {
-//            logger.error("Header login module needs SyncHandler. Can't get SyncHandler: '{}'", syncHandlerName);
-//            return;
-//        }
+        logger.debug("initialize called");
     }
 
     @Override
     public boolean abort() throws LoginException
     {
-        // TODO:  Do we need to override this method?
         logger.info("abort() called");
         return super.abort();
     }
@@ -183,7 +136,6 @@ public class Saml2LoginModule extends AbstractLoginModule {
     @Override
     protected void clearState()
     {
-        // TODO:  Do we need to override this method?
         logger.info("clearState() called");
         super.clearState();
     }
